updated dbb docs to use pyocd instead of hci
